+++ tensorflow/python/keras/layers/wrappers_test.py
@@ -732,6 +732,18 @@ class BidirectionalTest(test.TestCase):
       self.assertAllClose(y[0][0, 2], np.zeros(units * 2))
 
 
+  def test_TimeDistributed_learning_phase(self):
+    with self.cached_session():
+      # test layers that need learning_phase to be set
+      np.random.seed(1234)
+      x = keras.layers.Input(shape=(3, 2))
+      y = keras.layers.TimeDistributed(keras.layers.Dropout(.999))(
+          x, training=True)
+      model = keras.models.Model(x, y)
+      y = model.predict(np.random.random((10, 3, 2)))
+      self.assertAllClose(np.mean(y), 0., atol=1e-1, rtol=1e-1)
+
+
 def _to_list(ls):
   if isinstance(ls, list):
     return ls
